Move cli execute functions to controller
